pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "zero960"
version = "0.1.0"
description = "OpenEnv-oriented Chess960 self-improvement environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "chess>=1.11.2",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pydantic>=2.11.0",
  "uvicorn>=0.34.0",
]

[project.optional-dependencies]
train = [
  "torch>=2.6.0",
  "transformers>=4.52.0",
  "trl>=0.19.0",
]
openenv = [
  "openenv-core[core]>=0.2.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/zero960", "src/zero960_env"]

